Two bug fixes when maxVideoDuration option is set.
When you set the maxVideoDuration, like 30 seconds, it works on initial loading. But if you change album and go back to the camera roll, you can see the videos that are longer than 30 seconds.

Also, the number of media of the album does not match with the exact number of media when maxVideoDuration is set.